+++ Application/routes.py
@@ -1,29 +1,27 @@
 import datetime
 
-from flask import session, request
+from flask import request, Blueprint, session
 
-from Application.Model.Patients import Patients
-from Application.app import create_app, database
-from datetime import timedelta
+from Application.app import app
+from Application.database import database
 from Application.Model.Appointments import Appointments
 from Application.Model.Response import Response
 from Application.Model.Metadata import Metadata
 from flask import jsonify
-from flask_cors import CORS, cross_origin
+from flask_cors import cross_origin
 from Application.Utils.appointments_query_utils import query_field_parameters, search_fields, determine_sort_field, paginate, get_total_of_pages
-
-app = create_app()
-cors = CORS(app)
-app.config['CORS_HEADERS'] = 'Content-Type'
 
 @app.before_first_request
 def session_handler():
     database.create_all()
     session.permanent = True
-    app.permanent_session_lifetime = timedelta(minutes=30)
+    app.permanent_session_lifetime = datetime.timedelta(minutes=30)
+
+
+app1 = Blueprint('app1', __name__)
 
 @cross_origin()
-@app.route('/appointments/', methods=["GET", "POST", "PUT"])
+@app1.route('/appointments/', methods=["GET", "POST", "PUT"])
 def index():
     if request.method == "GET":
         query = request.query_string.decode()
@@ -46,7 +44,7 @@ def index():
         return jsonify(request.json), 200
 
 @cross_origin()
-@app.route('/appointments/<id>', methods=["DELETE", "PUT"])
+@app1.route('/appointments/<id>', methods=["DELETE", "PUT"])
 def index2(id):
     if request.method == "DELETE":
         Appointments.query.filter_by(id_appointment=int(id)).delete()
@@ -65,31 +63,3 @@ def index2(id):
         database.session.commit()
         return jsonify(request.json), 200
 
-
-@app.route('/patient/<cnp>', methods=["GET", "PUT"])
-def patient_home(cnp):
-    if request.method == "GET":
-        patient = Patients.query.filter_by(cnp_patient=cnp).first()
-
-        return jsonify(patient.serialize())
-
-    elif request.method == "PUT":
-        new_patient = request.json
-
-        patient = Patients.query.filter_by(cnp_patient=cnp).first()
-        patient.email = new_patient["email"]
-        patient.password = new_patient["password"]
-        patient.first_name = new_patient["first_name"]
-        patient.last_name = new_patient["last_name"]
-        patient.address = new_patient["address"]
-        patient.city = new_patient["city"]
-        patient.county = new_patient["county"]
-        patient.country = new_patient["country"]
-        patient.date_of_birth = datetime.datetime.strptime(new_patient['date_of_birth'], '%Y-%m-%d').date()
-
-        database.session.commit()
-        return jsonify(request.json), 200
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
